+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### Added 
+
+- Added `compact` parameter to `MaskedInput` https://github.com/Textualize/textual/pull/5952
+
 ### Fixed
 
 - Fixed `query_one` and `query_exactly_one` not raising documented `WrongType` exception.